Allow triangle slivers
A zero integer height does not mean no height. Including 0 (by comparing integer to < 1) was causing missing triangles at the far horizon of a flat ground plane. The integer clipping should only trigger if the triangle (as converted to row scanlines) is entirely past the bottom of the drawing scissor clipping area.